incorporated partial review 1
