+++ MongoDB.Benchmarks/MultiDoc/GridFSUploadBenchmark.cs
@@ -1,0 +1,62 @@
+/* Copyright 2021-present MongoDB Inc.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+using System;
+using System.IO;
+using MongoDB.Driver;
+using MongoDB.Driver.GridFS;
+using BenchmarkDotNet.Attributes;
+
+
+namespace MongoDB.Benchmarks.MultiDoc
+{
+    [IterationCount(100)]
+    [BenchmarkCategory("MultiBench", "WriteBench", "DriverBench")]
+    public class GridFsUploadBenchmark
+    {
+        private byte[] _fileBytes;
+        private MongoClient _client;
+        private GridFSBucket _gridFsBucket;
+
+        [GlobalSetup]
+        public void Setup()
+        {
+            string mongoUri = Environment.GetEnvironmentVariable("MONGODB_URI");
+            _client = mongoUri != null ? new MongoClient(mongoUri) : new MongoClient();
+            _client.DropDatabase("perftest");
+            _fileBytes = File.ReadAllBytes("../../../../../../../data/single_and_multi_document/gridfs_large.bin");
+            _gridFsBucket = new GridFSBucket(_client.GetDatabase("perftest"));
+        }
+
+        [IterationSetup]
+        public void BeforeTask()
+        {
+            _gridFsBucket.Drop();
+            _gridFsBucket.UploadFromBytes("smallfile", new byte[1]);
+        }
+
+        [Benchmark]
+        public void GridFsUpload()
+        {
+            _gridFsBucket.UploadFromBytes("gridfstest", _fileBytes);
+        }
+
+        [GlobalCleanup]
+        public void Teardown()
+        {
+            _client.DropDatabase("perftest");
+        }
+    }
+}
