+++ MongoDB.Benchmarks/SingleDoc/RunCommandBenchmark.cs
@@ -1,0 +1,46 @@
+/* Copyright 2021-present MongoDB Inc.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
+
+using System;
+using MongoDB.Bson;
+using MongoDB.Driver;
+using BenchmarkDotNet.Attributes;
+
+namespace MongoDB.Benchmarks.SingleDoc
+{
+    [IterationTime(3000)]
+    public class RunCommandBenchmark
+    {
+        private MongoClient _client;
+        private IMongoDatabase _database;
+
+        [GlobalSetup]
+        public void Setup()
+        {
+            string mongoUri = Environment.GetEnvironmentVariable("MONGODB_URI");
+            _client = mongoUri != null ? new MongoClient(mongoUri) : new MongoClient();
+            _database = _client.GetDatabase("admin");
+        }
+
+        [Benchmark]
+        public void RunCommand()
+        {
+            for (int i = 0; i < 10000; i++)
+            {
+                _database.RunCommand<BsonDocument>(new BsonDocument("hello", true));
+            }
+        }
+    }
+}
