--- benchmarks/benchmarks.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-  <PropertyGroup>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <DebugType>pdbonly</DebugType>
-    <DebugSymbols>true</DebugSymbols>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <Optimize>true</Optimize>
-    <Configuration>Release</Configuration>
-    <NoWarn>NU1701</NoWarn>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.8" />
-    <PackageReference Include="BenchmarkDotNet.Diagnostics.Windows" Version="0.13.8" Condition="'$(OS)' == 'Windows_NT'"/>
-    <PackageReference Include="NDesk.Options" Version="0.2.1" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\src\MongoDB.Driver.GridFS\MongoDB.Driver.GridFS.csproj" />
-    <ProjectReference Include="..\src\MongoDB.Driver\MongoDB.Driver.csproj" />
-  </ItemGroup>
-</Project>
